Add hook for storing and sharing reading context
This changes uses ReactHooks to provide a way for us to keep a global context for readings and pass it to whatever component that we want. We pass our reading generator to the context which sets the state whenever it needs to. Wherever we use our hook, it will re-render whenever the underlyign value changes