Add Metric Alarm components on the alarms page (#36)
This change updates the components for set parameters on the alarms screen to follow the specified ones in #20 : Showing upper and lower limits, the set value as well as the current value. The current value turns red if it breaches the limit. Moreover, the new components have been implemented with the user of styled components to improve readability and ease of styling. The styled components have also been used to make a global `Row` component to be used for other pages if needed.