Convert serial data values to SetParameter interfaces (#41)
The `SetParameter` interface was added to allow for adding of set values, upper limits and lower limits for parameters to allow for raising of alarms. This change calculates all these values using the default alarm thresholds and passes them down to the screens. These will be used to display a details parameter view on the alarms screen. 

For #9 and #20