Integrate readings into alarms (#42)
This change updates the `AlarmsScreen` component to make use of the `useReading` hook and construct alarm components based on the current parameter alarms we have available and update as the values change. To help with this, changes were made to `SerialParser` to calculate correct values and also set floor and ceilings to lower and upper limits where the value might be a lenghty decimal. Lastly, some adjustments were to the styles to allow for bigger value displays in the alarm component.

#20