Add status based alarms (#44)
This change adds the remaining alarms which are status based (can either be true or false rather than metric based). These have been added as a separate header to show these statues on the top with green values denoting normal status and red status with icon change to show alarm status. This completes the development for the alarms tab.

Resolves: #20