Refactor alarms with new ones (#54)
This change adds a new alarm at the end for `Low Respiratory Rate` as well as replacing the `Ventilation Active at Startup` to be a `System Reset` one. Since the latter does not have an icon in the connections header, a new icon set was added. Lastly, since the message is longer and not suitable for display, the logic for checking alarm eligibility was changed so that as long as our text is contained in the overall alarms text, it will be changed.