Update FiO2 Parameter Mechanism (#57)
The FiO2 parameter has changed to show a set value as a range from lower bound to upper bound. Additionally, the threshold values have also changed relative to these. This change updates the code to work with the change in spec. To do this, a new `setValueText` field has been added to the `SetParameter` interface to allow for setting a custom value for display. For showing on the monitoring panel, the already created `MetricDisplayString` component was used with the `setValueText` passed in. But for the alarms screen, functionality in `MetricCard` was extended to first check the `setValueText` value - if it is undefined, we just display the regular value instead.