Fix sweeping trends graphs (#71)
This change updates the logic for the graphs to properly follow the sweeping trends. This is done by using nulls instead of 0s to denote gaps or missing values. Moreover, the logic for calculating and adding gaps for sweeping trends has been update so that it scales with the amount of values we show on the graph.

Close #60