Add markers on graph to denote breath intake (#73)
This change monitors the breathing phase whenever data is parsed. If it detects the start of a new breath, it marks the data point for the graph and passes it along. In the `Graph` component, we have added a `Marker` component, which will create a triangle at the said location, below the `0` value in the graph. Once our graph is overwritten with new values, the markers from those position will be removed.

Close #61