Show blank tidal volume set point on pressure control modes (#89)
When our ventilation mode is pressure controlled i.e. `PCV` or `AC-PCV`, we need to show a blank value for our tidal volume set point in our monitoring display. To best handle this, a new component, `TidalVolumeMetricDisplay` was created, to encapsulate this ventilation check logic, taking in the ventilation mode and tidal volume parameter,setting the paramter set value to null in case of the relevant ventilation modes. In our `MetricDisplay` component, we handle null values by showng a `-`.

Close #81