Disable snoozing of Peak Pressure Alarm (#91)
Some alarms cannot have even their sound snoozed. One such is the `Peak Pressure` alarm. This change adds a quick check for the existence in current alarms to disallow even sound being snoozed for it. There is more work to be done, but that will be carried out in a separate refactor change.

Close #91