Implement smart management of alarms (#93)
We currently compare our newly received alarms with the previous alarms and decide if there is a change is alarms or not before showing them. However, in some cases, we may have already snoozed an alarm, and the change in alarms is only a decrease in the already existing ones. In such a case, we probably do not want to play our alarm sound. Therefore, some methods have been added to detect these cases. In case of high priority alarms, this does not apply so that edge case has also been added. Moreover, to improve readability, some refactoring has been applied.

Close #79