Fix FIO2 Display on Monitoring Panel (#94)
We were using the `MetricDisplayString` component for it before since we had to show a range value like `x-y %`. Now that it used actual value, this could had a lot of decimal values which wasn't great to look at from the UI. Converting it to `MetricDisplay` allows us to show it to rounded number like the rest.