Fix app crash when switching to Alarms Panel in Pressure Modes (#97)
The change to show a blank tidal volume set point in pressure mode ventilation mode in #89 was erroneously actually modifying the `setValue` field in the `setParameter` property passed down to it. This meant when we switch to the alarms panel, it would crash as the metric card component never expects a measured value to be `null` ever. This fix updates the code to not directly modify the property but instead pass the `null` forward to `MetricDisplay` in the `TidalVolumeMetricDisplay` and leave it unchanged. A little bit of refactoring was also carried out for escapsulation the ventilation mode check. Lastly, the version has been updated for this to be a patch fix.