Create multiple components for data management (#106)
The previous state of the application had a lot more coupling of different components, which made it difficult to introduce new behaviour as well as minimise impact of breaking changes. To mitigate this, the components have been modularised to a great extent. Now, all the data management is being done by `DataOrchestrator`. It is in charge of all data management: serial data packet retrieval, its parsing and logging data etc. Moreover, the update of the user interface has been moved to a new component called `UserInterfaceUpdater`, which now decides when to update the UI. Lastly, the serial data parsing is now a more pure function rather than having random dependencies. This `DataOrchestrator` also makes it easy to use the dummy data or the actual serial retriever, to allow for easier debugging. 

Close #74