Refactor workflows (#114)
This change refactors the workflows by converting into two separate ones. The `test` workflow runs on pull requests to `master` as well as any `push` to `master` as well. It is now also able to run in parallel with the `build-android` workflow as building the apk is separate from testing. Lastly, a badge has been added for showing the status for the tests.