Add UI for configurations screen (#118)
This change adds the user interface for allowing editing of ventilator configuration from this application. This is in the form of a picker for each configuration, allowing users to select from a bunch of options which have been hardcoded. Moreover, some logic has been added to add steps in enabling the settings to be edited and confirmed so that values are not accidentally changed. Lastly, a button is also added to allow users to start and stop ventilation.

Close #7