Prevent session data deletions (#124)
We stored our session data in an internal app directory. This meant that upon app deletion, all data in this directory was also removed, causing us to lose precious session data. To prevent this, we now store the session data in an external storage directory, provided we have been granted access to external storage. This change adds new utils to be able to do that as well as ensure consistency in the timestamped files such that all files will have names with the same timestamp, to allow for co-relation of the same session. 

The app logs have not been moved in the external storage directory for now as they are more specific to the running of the app and it requires adding a wrapper around the existing log library to handle our outstanding promises for app directory creation. As that is not too urgent, it has been left for the future. 

Close #119